fix: circular dependencies and make it work
