+++ control.lua
@@ -13,6 +13,7 @@ local Consumption = require("consumption")
 local City = require("city")
 local Queue = require("queue")
 local PrimaryIndustries = require("primary-industries")
+local UsedBottlesStore = require("used-bottles-store")
 
 --- TICK HANDLERS
 local ONE_SECOND = 60;
@@ -84,7 +85,7 @@ end
 
 script.on_nth_tick(ONE_MINUTE, function()
     for _, city in pairs(global.tycoon_cities or {}) do
-        UsedBottlesStore.return_used_bottles_to_market(city.id)
+        UsedBottlesStore.return_used_bottles_to_market(city)
     end
     consume_resources()
     City.construct_gardens()
